Use floor instead of round for decimal border widths
Chromium started to support subpixel borders 3 years ago
(see https://codereview.chromium.org/2640143005).
At that time it was decided to round decimal values
(as Edge was also doing it).
But edge rounded decimal values for both borders and outlines,
however Chromium only did it for borders (while flooring outlines).
Now that Edge uses Chromium, it has the same inconsistency too.

The other browsers Firefox and WebKit floor both borders and outlines,
so this patch is moving towards improving interoperability
between browsers on border pixel snapping.
This behavior is not defined in the spec, but if all browsers
behave the same that would simplify web author's lives.

The patch is basically changing roundf() calls by floorf(),
and then updating and rebaselining tests according to the new behavior.
It also adds two WPT tests to ensure that pixel snapping works
the same in all browsers.

BUG=1081246
TEST=fast/sub-pixel/sub-pixel-border-2.html
TEST=external/wpt/css/css-backgrounds/border-width-pixel-snapping-001-a.html
TEST=external/wpt/css/css-backgrounds/border-width-pixel-snapping-002-a.html

Change-Id: I14adfc78d2c9b7df697c7ce041dcbf1fe869e652