bluetooth: implement writeValueWith[out]Response()
This implements the Web Bluetooth BluetoothRemoteGATTCharacteristic
writeValueWithResponse() and writeValueWithoutResponse() methods.

https://webbluetoothcg.github.io/web-bluetooth/#writecharacteristicvalue

The writeValue() method is now deprecated and should continue to work
exactly as it did before.

New web_tests are just a copy of existing tests for writeValue(). New
non-generated tests have been converted to async/await. The tests for
writeValueWithoutResponse() have calls to setNextWriteResponse()
removed since they do not receive a response.

The getLastWrittenValue() method is modified to also return the write
type of the last written value. This is used to confirm that the correct
write type is requested.

Bug: 672648
Change-Id: I0d8032fded0028eaa78a55afe3336f253c85c590